Added collision detection, along with minor fixes
There's a problem with the gravity where when one mass approaches the x or y value of another, the force applied to it becomes overwhelmingly large and the smaller planet will shoot off. This is because velocity is handled by simply skipping over a larger distance in each frame, so the mass skips over a huge area and the force doesn't get rebalanced like it would in real life... I've tried applying max and min values to force, acceleration, velocity, and distance with no positive effects.